// src/app/models/pack-instruction-material.ts
export interface PackInstructionMaterial {
  packInstructionNumber?: null;
  packInstructionItem: number;
  packInstructionKey: string;
  layerTotal: number;
  parent: string;
  materialNumber: string;
  oldMaterialNumber: string;
  quantityPerLayer: number;
  baseUom: string;
  grossWeight: number;
  weightUom: string;
  volume: number;
  volumeUom: string;
  partQuantity: number;
  numberOfLayers: number;
  containsPart: boolean;
  layerVariable: boolean;
  iconType: string;
  edit: boolean;
  itemType: string;
  outErp: boolean;
  thuFlag: boolean;
  layerEnabled: boolean;
  quantityEnabled: boolean;
  length: number;
  width: number;
  height: number;
  dimensionUom: string;
  subHuIndicator: boolean;
  locked: boolean;
  itemNumberOfSoDocument: number;
  materialDescription: string;
  thuId: string;
  unitOfMeasure: string;
  key: string;
  iconName: string;
  imageUri: string;
  accountNumber: string;
  materialType: string;
  customerNumber: string;
  partNumber: string;
  originalNumberOfLayers: number;
  displayMode: boolean;
  typeText: string;
  dynamicCallOff: boolean;
  flexibleUnitLoad: boolean;
  mode: string;
}